backend/requirements.txt
fastapi
uvicorn
torch
torchvision
diffusers
transformers
pillow
opencv-python
numpy
